add lab-0

+++ Isu/Services/IIsuService.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Numerics;
 using Isu.Classes;
 
 namespace Isu.Services
@@ -8,7 +9,7 @@ namespace Isu.Services
         Group AddGroup(string name);
         Student AddStudent(Group group, string name);
 
-        Student GetStudent(int id);
+        Student GetStudent(BigInteger id);
         Student FindStudent(string name);
         List<Student> FindStudents(string groupName);
